+++ .eslintrc.yml
@@ -1,6 +1,8 @@
+parserOptions:
+  sourceType: 'module'
+
 env:
   es6: true
   browser: true
-  commonjs: true
 
 extends: 'htmlacademy/es6'
