+++ public/index.html
@@ -32,7 +32,7 @@
         <!-- Количество фильмов -->
       </section>
     </footer>
-    <script src="../src/main.js">
+    <script src="./bundle.js">
 
     </script>
   </body>
