+++ src/components/consts.js
@@ -1,0 +1,8 @@
+export const CARDS_AMOUNT = 5;
+export const CARDS_AMOUNT_EXTRA = 2;
+
+export const Position = {
+  BEFOREEND: `beforeend`,
+  AFTEREND: `afterend`,
+  AFTERBEGIN: `afterbegin`
+};
