+++ src/components/film-block.js
@@ -1,0 +1,3 @@
+const createFilmBlock = () => `<section class="films"></section>`;
+
+export default createFilmBlock;
