+++ src/components/more-button.js
@@ -1,0 +1,3 @@
+const createMoreButton = () => `<button class="films-list__show-more">Show more</button>`;
+
+export default createMoreButton;
